requirement.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
mysql==0.0.3
mysqlclient==2.2.0
phonenumbers==8.13.17
Pillow==10.0.0
pkg_resources==0.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
